Move the details of completion-resolution logic into Project
